add conftest file
